moving loopback tests into own file (external is for testing against an external implementation)
