+++ tests/external.rs
@@ -4,11 +4,10 @@ extern crate rand;
 extern crate xmodem;
 
 use std::process::{Command, Stdio, ChildStdin, ChildStdout};
-use std::io::{self, Read, Write, Seek, ErrorKind};
+use std::io::{self, Read, Write, Seek};
 use tempfile::NamedTempFile;
 use rand::{Rng, thread_rng};
 use xmodem::{Xmodem,Checksum};
-use std::sync::mpsc::{channel,Sender,Receiver};
 
 struct ChildStdInOut {
     stdin: ChildStdin,
@@ -29,40 +28,6 @@ impl Write for ChildStdInOut {
     fn flush(&mut self) -> io::Result<()> {
         self.stdin.flush()
     }
-}
-
-struct BidirectionalPipe {
-    pin : Receiver<u8>,
-    pout : Sender<u8>,
-}
-
-impl Read for BidirectionalPipe {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        for idx in 0..buf.len() {
-            buf[idx] = match self.pin.recv() {
-                Ok(v) => v,
-                Err(e) => return Err(std::io::Error::new(ErrorKind::BrokenPipe,e)),
-            }
-        }
-        Ok(buf.len())
-    }
-}
-
-impl Write for BidirectionalPipe {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        for v in buf { self.pout.send(*v).unwrap(); }
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
-fn loopback() -> (BidirectionalPipe, BidirectionalPipe) {
-    let (s1,r1) = channel();
-    let (s2,r2) = channel();
-    (BidirectionalPipe{ pin : r1, pout : s2 }, BidirectionalPipe{ pin : r2, pout : s1 })
 }
 
 #[cfg(test)]
@@ -97,48 +62,6 @@ fn xmodem_recv(checksum_mode : Checksum,data_len : usize) {
          padded_data.push(0x1a);
     }
     assert_eq!(padded_data, recv_data);
-}
-
-#[cfg(test)]
-fn xmodem_loopback(checksum_mode:Checksum, data_len : usize) {
-    let mut data_out = vec![0; data_len];
-    // We don't really need the rng here
-    for idx in 0..data_len { data_out[idx] = ((idx+7) * 13) as u8; }
-    let (mut p1, mut p2) = loopback();
-    let handle = std::thread::spawn(move || {
-        let mut xmodem = Xmodem::new();
-        xmodem.send(&mut p1, &mut &data_out[..]).unwrap();
-        data_out
-    });
-    let handle2 = std::thread::spawn(move || {
-        let mut xmodem = Xmodem::new();
-        let mut data_in= vec![0; 0];
-        xmodem.recv(&mut p2, &mut data_in, checksum_mode).unwrap();
-        data_in
-    });
-    
-    let mut dato = handle.join().unwrap();
-    // Pad output data to multiple of 128 for comparison
-    for _ in 0..(128 - data_len % 128) { dato.push(0x1a); }
-    let dati = handle2.join().unwrap();
-    assert_eq!(dato.len(),dati.len());
-    assert_eq!(dato,dati);
-}
-
-#[test]
-fn xmodem_loopback_standard() {
-    xmodem_loopback(Checksum::Standard,2000);
-}
-
-#[test]
-fn xmodem_loopback_crc() {
-    xmodem_loopback(Checksum::CRC16,2000);
-}
-
-#[test]
-fn xmodem_loopback_long_crc() {
-    // make sure we wrap block counter
-    xmodem_loopback(Checksum::CRC16,50000);
 }
 
 #[test]
